backend/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.29.0
django-filter==24.3
django-cors-headers==4.4.0
django-redis==5.4.0
channels==4.3.2
channels-redis==4.3.0
daphne==4.1.2
uvicorn==0.32.0
gunicorn==21.2.0
celery==5.4.0
redis==5.0.1
docker==7.1.0
boto3==1.42.2
botocore==1.42.2
django-simple-captcha==0.6.0
Pillow==11.0.0
